requirements.txt
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
deepspeed>=0.14.0
triton>=2.1.0
wandb>=0.16.0
einops>=0.7.0
datasets>=2.18.0
opencv-python>=4.9.0
decord>=0.6.0
pyyaml>=6.0
tqdm>=4.66.0
numpy<2.0.0,>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
qwen-vl-utils>=0.0.1
pillow>=10.0.0
gputil>=1.4.0
psutil>=5.9.0
